# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nextstep Dashboard API
  version: 1.1.0
  description: >-
    Contract for the multi-tenant SML dashboard, production administration,
    distributed report execution, LINE delivery, and LIFF viewer. Mutation
    requests authenticated by cookie also require X-CSRF-Token.
servers:
  - url: https://dashboard.nextstep-soft.com
...
paths:
...
  /api/v1/admin/report-runs:
    get:
      tags: [Admin operations]
      operationId: listAdminReportRuns
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: tenantId, in: query, schema: {type: string, format: uuid}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ReportRunStatus'}}
      responses:
        '200':
          description: Report run history.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportRunPage'}
  /api/v1/admin/line-quota:
    get:
      tags: [Admin operations]
      operationId: getLineQuota
      description: Provider-wide shared OA quota status. Consumption includes messages sent by other systems and LINE Official Account Manager.
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Safe quota metadata without channel credentials.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LineQuotaStatus'}
...
  /api/v1/admin/audit-logs:
    get:
      tags: [Admin operations]
      operationId: listAuditLogs
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: tenantId, in: query, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Redacted audit events.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: nextstep_admin_session
    ViewerSession:
      type: apiKey
      in: cookie
      name: nextstep_viewer_session
  parameters:
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
  schemas:
...
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: string, nullable: true}
        hasMore: {type: boolean}
...
    ReportKey:
      type: string
      enum: [sales_goods_services, purchase_goods_payables, gross_profit_by_product, gross_profit_by_ar_customer, stock_balance, stock_reorder, ar_customer_movement, ar_debt_receipt, cash_bank_receipts, cash_bank_payments]
...
    ReportRunStatus: {type: string, enum: [QUEUED, CLAIMED, RUNNING, SUCCEEDED, FAILED, CANCELLED, EXPIRED]}
...
    ReportRun:
      type: object
      required: [id, tenantId, reportKey, status, periodPreset, rowCount, isTruncated, queuedAt, expiresAt]
      properties:
        id: {type: string, format: uuid}
        tenantId: {type: string, format: uuid}
        tenantName: {type: string, description: Present on Admin report history responses.}
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        status: {$ref: '#/components/schemas/ReportRunStatus'}
        runtimeStatus: {type: string, enum: [ACTIVE, STALLED], description: Operational lease health for Admin monitoring.}
        retryAvailableAt: {type: string, format: date-time, nullable: true}
        waitReason: {type: string, enum: [TENANT_BUSY, HOST_BUSY, TENANT_COOLDOWN, HOST_COOLDOWN, SCHEDULE_RESERVED]}
        resultKind: {type: string, enum: [SUMMARY, DETAIL]}
        periodPreset: {type: string}
        dateFrom: {type: string, format: date, nullable: true}
        dateTo: {type: string, format: date, nullable: true}
        rowCount: {type: integer, minimum: 0}
        isTruncated: {type: boolean}
        summary: {type: object, additionalProperties: true}
        reconciliation: {type: object, additionalProperties: true}
        safeErrorCode: {type: string, nullable: true}
        safeErrorMessage: {type: string, nullable: true}
        queuePosition: {type: integer, minimum: 1, nullable: true}
        queuedAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        finishedAt: {type: string, format: date-time, nullable: true}
        expiresAt: {type: string, format: date-time}
        progress: {$ref: '#/components/schemas/ReportRunProgress'}
    ReportRunProgress:
      type: object
      additionalProperties: false
      required: [phase, phaseSequence, completedSteps, totalSteps, attempt, sampleCount, elapsedMs]
      properties:
        phase: {type: string, enum: [QUEUED, CONNECTING, QUERYING_CURRENT, QUERYING_COMPARISON, BUILDING_DASHBOARD, SAVING_RESULT, WAITING_RETRY, COMPLETED]}
        phaseSequence: {type: integer, minimum: 0}
        completedSteps: {type: integer, minimum: 0}
        totalSteps: {type: integer, minimum: 0}
        attempt: {type: integer, minimum: 0}
        updatedAt: {type: string, format: date-time, nullable: true}
        expectedP50Ms: {type: integer, minimum: 1, nullable: true}
        expectedP90Ms: {type: integer, minimum: 1, nullable: true}
        sampleCount: {type: integer, minimum: 0}
        elapsedMs: {type: integer, minimum: 0}
    ReportRunPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/ReportRun'}}
        page: {$ref: '#/components/schemas/PageInfo'}
...
    LineQuotaStatus:
      type: object
      additionalProperties: false
      required: [state, providerLimit, providerConsumed, locallyAccepted, operationalReservePercent, syncedAt]
      properties:
        state: {type: string, enum: [READY, UNLIMITED, STALE, UNSYNCED]}
        providerLimit: {type: integer, minimum: 0, nullable: true, description: "Provider target limit; null means unlimited or not yet synced, distinguished by state."}
        providerConsumed: {type: integer, minimum: 0, nullable: true, description: Approximate provider-wide consumption including Nexflow and OA Manager.}
        locallyAccepted: {type: integer, minimum: 0, description: Push messages accepted from Nextstep Dashboard in the current month.}
        operationalReservePercent: {type: integer, minimum: 0, maximum: 50}
        syncedAt: {type: string, format: date-time, nullable: true}
...
    AuditEvent:
      type: object
      required: [id, actorType, action, resourceType, result, createdAt]
      properties:
        id: {type: string, format: uuid}
        tenantId: {type: string, format: uuid, nullable: true}
        tenantName: {type: string, nullable: true}
        actorType: {type: string, enum: [ADMIN, VIEWER, WORKER, SYSTEM]}
        action: {type: string}
        resourceType: {type: string}
        resourceId: {type: string, nullable: true}
        result: {type: string, enum: [SUCCESS, DENIED, FAILED]}
        safeErrorCode: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
    AuditPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/AuditEvent'}}
        page: {$ref: '#/components/schemas/PageInfo'}